Separate Django App for homepage
