+++ pontoon/homepage/static/img/moz-logo.svgz
