+++ pontoon/homepage/static/js/home.js
