+++ pontoon/homepage/templates/home.html
@@ -24,11 +24,11 @@
         <div class="title-style">A place to localize Mozilla, fast and easy.</div>
         <div class="content-style">Join a global community of thousands in making Firefox and various Mozilla products speak to over a hundred of cultures in their native language.</div>
         <div class="flex vertical-align-portrait">
-          <a class="homepage-button primary-button-cta" href="/teams/">
+          <a class="homepage-button primary-button-cta" href="{{ url('pontoon.teams') }}">
             <div>Start Localizing Now</div>
           </a>
           <p>or</p>
-          <a class="homepage-button secondary-button-cta" href="{{ url('pontoon.translate.locale.agnostic', 'projects', 'demo', 'all-resources') }}">
+          <a class="homepage-button secondary-button-cta" href="/tour">
             <div>Take a Tour</div>
           </a>
         </div>
@@ -39,7 +39,7 @@
     <!-- Section-2: Why Mozilla? -->
     <section class="snaps section-2 white-background" data-section-name="section-2">
       <div class="container vertical-align-center container--full-height flex--wrap">
-        <div  class= "flex-col-full flex-col-5"><img class="section-side-image-1 neg-margin--large" src="/static/img/Homepage/section-2.svg" /></div>
+        <div  class= "flex-col-full flex-col-5"><img class="section-side-image-1 neg-margin--large" src="/static/img/section-2.svg" /></div>
         <div class="why-mozilla flex-col-full flex-col-7 flex-direction-col neg-margin--large">
           <div class="title-style colored-title-style">Here's why localizing Mozilla matters.</div>
           <div class="content-style colored-content-style">Mozilla's localization (L10n) objective is to improve the world by culturally adapting Mozilla products by region and offering them to every user throughout the world and create an open web available for everyone beyond all the linguistic, cultural and geographical boundaries.<br/><br/>This program is a non-profit open source project where we work closely with communities of volunteer contributors who also care about the fate of the open web. Their contributions to our L10n effort make having an open and accessible web possible.<br/><br/><b>Working together, we can open the web to all and protect user rights all over the world.</b>
@@ -60,7 +60,7 @@
               <div class="text-3"><p>Suggestion gets approved by a trusted localizer</p></div>
               <div class="text-4 align-right"><p>And it lands in product</p></div>
             </div>
-            <img class="section-side-image-2" src="/static/img/Homepage/section-3.svg" />
+            <img class="section-side-image-2" src="/static/img/section-3.svg" />
           </div>
           <div class="colored-content-style align-center path-startend-text"><p>Show it to your friends</p></div>
         </div>
@@ -83,37 +83,37 @@
           </div>
           <div class="cards">
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-1.svg" />
+              <div class="card-image"><img src="/static/img/feature-1.svg" />
               </div>
               <div class="card-text content-style colored-content-style">Built in machine translation and translation memory services.
               </div>
             </div>
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-2.svg" />
+              <div class="card-image"><img src="/static/img/feature-2.svg" />
               </div>
               <div class="card-text content-style colored-content-style">Notifications to stay on top of localization all the time.
               </div>
             </div>
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-3.svg" />
+              <div class="card-image"><img src="/static/img/feature-3.svg" />
               </div>
               <div class="card-text content-style colored-content-style">Tracking progress, deadlines and priorities on dashboard.
               </div>
             </div>
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-4.svg" />
+              <div class="card-image"><img src="/static/img/feature-4.svg" />
               </div>
               <div class="card-text content-style colored-content-style">Mass actions for quick translation and review tasks.
               </div>
             </div>
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-5.svg" />
+              <div class="card-image"><img src="/static/img/feature-5.svg" />
               </div>
               <div class="card-text content-style colored-content-style">Automated bi-di synchronization with version control systems.
               </div>
             </div>
             <div class="card-features flex vertical-align-center">
-              <div class="card-image"><img src="/static/img/Homepage/feature-6.svg" />
+              <div class="card-image"><img src="/static/img/feature-6.svg" />
               </div>
               <div class="card-text content-style colored-content-style">In-context localization.
               </div>
@@ -133,7 +133,7 @@
             <div>Give it a Try!</div>
           </a>
         </div>
-        <div  class= "flex-col-full flex-col-6 align-right portrait-align-center"><img class="section-side-image-3 neg-margin--large" src="/static/img/Homepage/section-5.svg" /></div>
+        <div  class= "flex-col-full flex-col-6 align-right portrait-align-center"><img class="section-side-image-3 neg-margin--large" src="/static/img/section-5.svg" /></div>
       </div>
     </section>
 
@@ -142,19 +142,19 @@
       <div class="footer vertical-align-center flex-direction-col">
         <div class="title-style">Join us.</div>
         <p class="content-style">Now that you know what Pontoon is, come on in.</p>
-        <a class="homepage-button primary-button-cta" href="/teams/">
+        <a class="homepage-button primary-button-cta" href="{{ url('pontoon.teams') }}">
           <div>Start Localizing Now</div>
         </a>
         <p class="content-style">Want to learn more?</p>
-        <a class="homepage-button secondary-button-cta" href="{{ url('pontoon.translate.locale.agnostic', 'projects', 'demo', 'all-resources') }}">
+        <a class="homepage-button secondary-button-cta" href="/tour">
           <div>Take a Tour</div>
         </a>
       </div>
       <div class="homepage-footer flex">
-        <div class="flex-col-4 align-left"><a href="https://mozilla.org/"><img class="moz-logo" src="/static/img/Homepage/moz-logo.svgz" /></a></div>
+        <div class="flex-col-4 align-left"><a href="https://mozilla.org/"><img class="moz-logo" src="/static/img/moz-logo.svgz" /></a></div>
         <div class="flex-col-4 flex github--link">
           <p class="content-style"><a href="https://github.com/mozilla/pontoon/">Hack it on Github</a></p>
-          <img class="github" src="static/img/Homepage/section-6-github.svg" />
+          <img class="github" src="static/img/section-6-github.svg" />
         </div>
         <div class="flex-col-4 align-right">
           <p class="content-style"><a href="mailto:pontoon-team@mozilla.com">Get in touch</a></p>
