+++ pontoon/urls.py
@@ -113,6 +113,7 @@ urlpatterns = [
     url(r'', include('pontoon.base.urls')),
     url(r'', include('pontoon.batch.urls')),
     url(r'', include('pontoon.api.urls')),
+    url(r'', include('pontoon.homepage.urls')),
 
     # Team page: Must be at the end
     url(r'^(?P<locale>[A-Za-z0-9\-\@\.]+)/$', team, name='pontoon.teams.team'),
